- typecast malloc / Realloc issues.
- signed / unsigned issues.
(This used to be commit c8fd555179314baf1672a23db34dc8ad9f2d02bf)

+++ source3/lib/genrand.c
@@ -58,8 +58,8 @@ static void do_dirrand(char *name, unsigned char *buf, int buf_len)
 {
   DIR *dp = opendir(name);
   pstring fullname;
-  int len_left;
-  int fullname_len;
+  size_t len_left;
+  size_t fullname_len;
   char *pos;
 
   pstrcpy(fullname, name);
